refactor: split app navigation helpers
